+++ config.php
@@ -1,5 +1,9 @@
 <?php
 
 $config = [
-  'tpl_path' => 'templates/'
+  'tpl_path' => 'templates/',
+  'db_host' => 'localhost',
+  'db_user' => 'root',
+  'db_pass' => '',
+  'db_database' => 'yeticave'
 ];
